+++ docs/conf.py
@@ -355,6 +355,9 @@ needs_render_context = {
 # build needs.json to make permalinks work
 needs_build_json = True
 
+# build json file include needs with the same docs_name
+needs_per_page = True
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]