+++ sphinx_needs/needs.py
@@ -12,8 +12,10 @@ import sphinx_needs.debug as debug  # Need to set global var in it for timeing m
 from sphinx_needs.api.configuration import add_extra_option
 from sphinx_needs.builder import (
     NeedsBuilder,
+    NeedsPerPageBuilder,
     NeedumlsBuilder,
     build_needs_json,
+    build_needs_per_page_json,
     build_needumls_pumls,
 )
 from sphinx_needs.config import NEEDS_CONFIG
@@ -141,6 +143,7 @@ def setup(app: Sphinx) -> Dict[str, Any]:
 
     app.add_builder(NeedsBuilder)
     app.add_builder(NeedumlsBuilder)
+    app.add_builder(NeedsPerPageBuilder)
     app.add_config_value(
         "needs_types",
         [
@@ -281,6 +284,10 @@ def setup(app: Sphinx) -> Dict[str, Any]:
     #
     app.add_config_value("needs_debug_measurement", False, "html", types=[dict])
 
+    # add config for needs id with the sames docs_name
+    app.add_config_value("needs_per_page", False, "html", types=[bool])
+    app.add_config_value("needs_per_page_build_path", "need_per_page", "html")
+
     # Define nodes
     app.add_node(Need, html=(html_visit, html_depart), latex=(latex_visit, latex_depart))
     app.add_node(
@@ -376,6 +383,8 @@ def setup(app: Sphinx) -> Dict[str, Any]:
     app.connect("env-updated", install_lib_static_files)
     app.connect("env-updated", install_permalink_file)
 
+    #
+    app.connect("build-finished", build_needs_per_page_json)
     # This should be called last, so that need-styles can override styles from used libraries
     app.connect("env-updated", install_styles_static_files)